Hent antall sykmeldte med en enkelt query
NærmesteLederRepository og SykmeldingRepository (med tilhørende klasser) kan
slettes når frontend er migrert til nye versjonen av endepunktet.
