+++ src/main/java/no/nav/arbeidsgiver/min_side/services/digisyfo/SykmeldingRepositoryStub.java
@@ -1,11 +1,9 @@
 package no.nav.arbeidsgiver.min_side.services.digisyfo;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Map;
 
 @Profile({"local", "labs"})
@@ -15,19 +13,5 @@ public class SykmeldingRepositoryStub implements SykmeldingRepository {
     @Override
     public Map<String, Integer> oversiktSykmeldinger(String nærmestelederFnr) {
         return Map.of("910825526", 4);
-    }
-
-    @Override
-    public Map<String, Integer> oversiktSykmeldte(String nærmestelederFnr) {
-        return Map.of("910825526", 4);
-    }
-
-    @Override
-    public List<Virksomhetsinfo> sykmeldtePrVirksomhet(String nærmestelederFnr) {
-        return List.of(new Virksomhetsinfo("910825526", 4));
-    }
-
-    @Override
-    public void processEvent(List<ImmutablePair<String, SykmeldingHendelse>> records) {
     }
 }
